fix: do not show welcome modal if nickname exist
Reproduction:
This issue was happening when user logged in to the same hp on a different browser. That results in the previous session being invalidated. When user refreshes the app, it logs him out. Trying to log in on the SAME session (no refresh anymore) results in welcome modal being shown.